backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.17
python-dotenv==1.0.1
phonenumbers==8.13.45
python-dateutil==2.9.0.post0
tzdata==2024.1

